@@ -1,6 +1,7 @@
 package no.nav.pam.annonsemottak.markdown;
 
-import org.junit.Test;
+
+import org.junit.jupiter.api.Test;
 
 import static org.junit.Assert.assertEquals;
 
