@@ -1,7 +1,7 @@
 package no.nav.pam.annonsemottak.markdown;
 
-import org.apache.commons.text.StringEscapeUtils;
-import org.junit.Test;
+
+import org.junit.jupiter.api.Test;
 
 import static org.hamcrest.CoreMatchers.containsString;
 import static org.hamcrest.CoreMatchers.not;
