@@ -1,6 +1,6 @@
 package no.nav.pam.annonsemottak.markdown;
 
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 
 import static org.hamcrest.CoreMatchers.containsString;
 import static org.junit.Assert.assertEquals;
