@@ -2,15 +2,16 @@ package no.nav.pam.annonsemottak.receivers.dexi;
 
 import no.nav.pam.annonsemottak.Application;
 import no.nav.pam.annonsemottak.stilling.Stilling;
-import org.junit.AfterClass;
-import org.junit.BeforeClass;
-import org.junit.Test;
-import org.junit.runner.RunWith;
+
+import org.junit.jupiter.api.AfterAll;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.test.annotation.Rollback;
 import org.springframework.test.context.ContextConfiguration;
-import org.springframework.test.context.junit4.SpringRunner;
+import org.springframework.test.context.junit.jupiter.SpringExtension;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.io.IOException;
@@ -21,7 +22,7 @@ import java.util.stream.Collectors;
 
 import static no.nav.pam.annonsemottak.stilling.StillingTestdataBuilder.stilling;
 
-@RunWith(SpringRunner.class)
+@ExtendWith(SpringExtension.class)
 @DataJpaTest
 @Rollback
 @Transactional
@@ -53,12 +54,12 @@ public class DexioConnectorTestIT {
     private String tiltredelse = "Tiltredelse";
     private String stillingstype = "Type stilling";
 
-    @BeforeClass
+    @BeforeAll
     public static void beforeClass() {
         System.setProperty("dexi.url", "https://api.dexi.io");
     }
 
-    @AfterClass
+    @AfterAll
     public static void afterClass() {
         System.clearProperty("dexi.url");
     }
